requirements.txt
Flask
waitress
psycopg2-binary
Flask-HTTPAuth
python-dotenv